Tapo C320WS camera returns origin not according to SDP RFC
As a reply to the DESCRIBE method, this camera returned
an origin such as:

  o=- 14665860 31787219 1 IN IP4 192.168.0.155

This is not according to the RFC 4566, which states:

  o=<username> <sess-id> <sess-version> <nettype> <addrtype> <unicast-address>

So it seems like the '1' is the offending part in the origin string.

This ugly fix will first try to parse the origin string according to
the RFC and if that fails try to parse the faulty 7 element string.